Organised the project structure
