--- src/main/kotlin/com/ex_dock/ex_dock/StartVerticle.kt
@@ -1,6 +1,0 @@
-package com.ex_dock.ex_dock
-
-import io.vertx.core.AbstractVerticle
-
-class StartVerticle: AbstractVerticle() {
-}
