+++ src/main/kotlin/com/ex_dock/ex_dock/frontend/FrontendStarter.kt
@@ -1,0 +1,6 @@
+package com.ex_dock.ex_dock.frontend
+
+import io.vertx.core.AbstractVerticle
+
+class FrontendStarter: AbstractVerticle() {
+}
